Inspect variables

+++ src/main/java/org/javacs/debug/ContinuedEventBody.java
@@ -2,7 +2,7 @@ package org.javacs.debug;
 
 public class ContinuedEventBody {
     /** The thread which was continued. */
-    public int threadId;
+    public long threadId;
     /** If 'allThreadsContinued' is true, a debug adapter can announce that all threads have continued. */
     public Boolean allThreadsContinued;
 }
