+++ src/main/java/org/javacs/debug/GotoArguments.java
@@ -3,7 +3,7 @@ package org.javacs.debug;
 /** Arguments for 'goto' request. */
 public class GotoArguments {
     /** Set the goto target for this thread. */
-    public int threadId;
+    public long threadId;
     /** The location where the debuggee will continue to run. */
     public int targetId;
 }
