+++ src/main/java/org/javacs/debug/NextArguments.java
@@ -3,5 +3,5 @@ package org.javacs.debug;
 /** Arguments for 'next' request. */
 public class NextArguments {
     /** Execute 'next' for this thread. */
-    public int threadId;
+    public long threadId;
 }
