+++ src/main/java/org/javacs/debug/PauseArguments.java
@@ -3,5 +3,5 @@ package org.javacs.debug;
 /** Arguments for 'pause' request. */
 public class PauseArguments {
     /** Pause execution for this thread. */
-    public int threadId;
+    public long threadId;
 }
