+++ src/main/java/org/javacs/debug/RestartFrameArguments.java
@@ -3,5 +3,5 @@ package org.javacs.debug;
 /** Arguments for 'restartFrame' request. */
 public class RestartFrameArguments {
     /** Restart this stackframe. */
-    public int frameId;
+    public long frameId;
 }
