+++ src/main/java/org/javacs/debug/ScopesArguments.java
@@ -3,5 +3,5 @@ package org.javacs.debug;
 /** Arguments for 'scopes' request. */
 public class ScopesArguments {
     /** Retrieve the scopes for this stackframe. */
-    public int frameId;
+    public long frameId;
 }
