+++ src/main/java/org/javacs/debug/SetVariableArguments.java
@@ -3,7 +3,7 @@ package org.javacs.debug;
 /** Arguments for 'setVariable' request. */
 public class SetVariableArguments {
     /** The reference of the variable container. */
-    public int variablesReference;
+    public long variablesReference;
     /** The name of the variable in the container. */
     public String name;
     /** The value of the variable. */
