+++ src/main/java/org/javacs/debug/StepInTargetsArguments.java
@@ -3,5 +3,5 @@ package org.javacs.debug;
 /** Arguments for 'stepInTargets' request. */
 public class StepInTargetsArguments {
     /** The stack frame for which to retrieve the possible stepIn targets. */
-    public int frameId;
+    public long frameId;
 }
