+++ src/main/java/org/javacs/debug/ThreadEventBody.java
@@ -4,5 +4,5 @@ public class ThreadEventBody {
     /** The reason for the event. Values: 'started', 'exited', etc. */
     public String reason;
     /** The identifier of the thread. */
-    public int threadId;
+    public long threadId;
 }
